fix: Adjust how navigation containers are used so that ICCA still works
