+++ src/frontend/AppContainerWrapper.tsx
@@ -13,8 +13,8 @@ import SettingsContext from "./context/SettingsContext";
 // import useProjectInviteListener from "./hooks/useProjectInviteListener";
 import IS_E2E from "./lib/is-e2e";
 import bugsnag from "./lib/logger";
-import { AppStack } from "./NavigationStacks/AppStack";
-import { WithModalsStack } from "./NavigationStacks/WithModalsStack";
+import DefaultContainer from "./Navigation/DefaultContainer";
+import OnboardingContainer from "./Navigation/OnboardingContainer";
 
 // Turn on logging if in debug mode
 if (__DEV__) debug.enable("*");
@@ -133,11 +133,9 @@ const AppContainerWrapper = () => {
     [experiments.onboarding]
   );
 
-  const AppContainer = React.useMemo(
-    () =>
-      createAppContainer(experiments.onboarding ? WithModalsStack : AppStack),
-    [experiments.onboarding]
-  );
+  const AppContainer = experiments.onboarding
+    ? OnboardingContainer
+    : DefaultContainer;
 
   /**
    * TODO: Uncomment when project invites are supported
