+++ src/frontend/Navigation/DefaultContainer.ts
@@ -1,0 +1,4 @@
+import { AppStack } from "./AppStack";
+import { createAppContainer } from "react-navigation";
+
+export default createAppContainer(AppStack);
