+++ src/frontend/Navigation/OnboardingContainer.ts
@@ -1,0 +1,4 @@
+import { OnboardingStack } from "./OnboardingStack";
+import { createAppContainer } from "react-navigation";
+
+export default createAppContainer(OnboardingStack);
